In Application.cpp the 3Dconnexion.h file in inckuded and the init function is called initializing the device.
The native support for the 3Dconnexion device is added in the 3Dconnexion.cpp en 3Dconnexion.h files.

There is a connexionData class for the input mapper.  There is a connexionClient class wich is different for osx and windows. Windows has got a separate params class.
The osx connexionClient is more or less the official api implementation .
The win connexionClient ia  the unofficial way to implement. It reads the raw input usb data, The official api does not seem to work under windows with c++ and qt.