update buttons
Put the zooming again on button_1 and button_2
it is a hard to press both buttons at the same time, almost ever
button_1 of button_2 gets called
It is very hard to press a button and at the same time move or rotate
the device, so I removed the simultaneous button and rotation.

the head in now moved down or up with tilting forward or backward. It
is a bit hard to do this without moving the avatar a bit.

The buttons should now work on windows

reverted the adding of the framework since this does not work

and hopefully resolved the merge conflict.