fix maximo cache crash
Added a check if the fileName() checked is the same as the file in the
cache.
Move the refresh into the if, so the cache only gets refreshed when
needed.

The fst file is cached ok, and loaded almost instantly when in the
cache. However when the fst is loaded in, it causes the maybeRefresh
being called a second time with the fbx file, this should not trigger a
refresh of the fst file.